+++ Objects/FXDNumber.m
@@ -41,16 +41,16 @@
 	NSInteger fileSize = [self integerValue];
 
 	if (fileSize < kiloUnit) {
-		formattedString = [NSString stringWithFormat:@"%d B", fileSize];
+		formattedString = [NSString stringWithFormat:@"%ld B", (long)fileSize];
 	}
 	else if (fileSize < (kiloUnit*kiloUnit)) {
-		formattedString = [NSString stringWithFormat:@"%d KB", (fileSize /kiloUnit)];
+		formattedString = [NSString stringWithFormat:@"%ld KB", (long)(fileSize /kiloUnit)];
 	}
 	else if (fileSize < (kiloUnit*kiloUnit*kiloUnit)) {
-		formattedString = [NSString stringWithFormat:@"%d MB", (fileSize /(kiloUnit*kiloUnit))];
+		formattedString = [NSString stringWithFormat:@"%ld MB", (long)(fileSize /(kiloUnit*kiloUnit))];
 	}
 	else if (fileSize < (kiloUnit*kiloUnit*kiloUnit*kiloUnit)) {
-		formattedString = [NSString stringWithFormat:@"%d GB", (fileSize /(kiloUnit*kiloUnit*kiloUnit))];
+		formattedString = [NSString stringWithFormat:@"%ld GB", (long)(fileSize /(kiloUnit*kiloUnit*kiloUnit))];
 	}
 
 	return formattedString;
@@ -64,17 +64,17 @@
 	NSInteger seconds = (NSInteger)timerInterval % 60;
 
 	if (seconds < 10) {
-		formattedString = [NSString stringWithFormat:@"0%d", seconds];
+		formattedString = [NSString stringWithFormat:@"0%ld", (long)seconds];
 	}
 	else {
-		formattedString = [NSString stringWithFormat:@"%d", seconds];
+		formattedString = [NSString stringWithFormat:@"%ld", (long)seconds];
 	}
 
 
 	NSInteger minutes = (NSInteger)timerInterval / 60;
 
 	if (minutes < 60) {
-		formattedString = [NSString stringWithFormat:@"%d:%@", minutes, formattedString];
+		formattedString = [NSString stringWithFormat:@"%ld:%@", (long)minutes, formattedString];
 
 		return formattedString;
 	}
@@ -84,13 +84,13 @@
 	minutes = minutes % 60;
 
 	if (minutes < 10) {
-		formattedString = [NSString stringWithFormat:@"0%d:%@", minutes, formattedString];
+		formattedString = [NSString stringWithFormat:@"0%ld:%@", (long)minutes, formattedString];
 	}
 	else {
-		formattedString = [NSString stringWithFormat:@"%d:%@", minutes, formattedString];
+		formattedString = [NSString stringWithFormat:@"%ld:%@", (long)minutes, formattedString];
 	}
 
-	formattedString = [NSString stringWithFormat:@"%d:%@", hours, formattedString];
+	formattedString = [NSString stringWithFormat:@"%ld:%@", (long)hours, formattedString];
 	
 
 	return formattedString;
